[master] 2/7 2+2+
Description

It is high time to improve the previous version of the calculator. What if we have many pairs of numbers, the sum of which we need to find? It will be very inconvenient to run the program every time. So then, let's add a loop to calculate the sum of two numbers continuously. Be sure to have a safeword to break the loop. Also, It would be nice to think through situations where users enter only one number or do not enter numbers at all.

Objective

1. Write a program that reads two numbers in a loop and prints the sum in the standard output.
2. The program should print the same number if a user enters only a single number. If a user enters an empty line, the program should ignore it and continue the loop until the user enters a number.
3. When the command /exit is entered, the program must print "Bye!" (without quotes), and then stop.