[master] 3/7 Count them all
Description

In rare cases, we need to calculate the sum of only two numbers. Now it is time to teach the calculator to read an unlimited sequence of numbers. Also, let's take care of ourselves if after a while we want to remember what our program does. For this purpose, we'll introduce a new command /help to our calculator. Users who have first exposure to this program may use /help as well to know more about it!

Objective

1. Add to the calculator the ability to read an unlimited sequence of numbers.
2. Add a /help command to print some information about the program.
3. If you encounter an empty line, do not output anything.