[master] 4/7 Add subtractions
Description

Finally, we got to the next operation: subtraction. It means that from now on the program must receive the addition + and subtraction - operators as an input to distinguish operations from each other. It must support both unary and binary minus operators. Moreover, If the user has entered several same operators following each other, the program still should work. Also, as you remember from school math, two adjacent minus signs turn into a plus. The smart calculator ought to have such a feature.

Pay attention to the /help command, it is important to maintain its relevance depending on the changes (in the next stages too). You can write information about your program in free form, but the main thing is that it should be understandable to you and other users.

Objective

1. The program must calculate expressions like these: 4 + 6 - 8, 2 - 3 - 4, and so on.
2. Modify the result of the /help command to explain these operations.
3. Decompose your program using functions to make it easy to understand and edit later.
4. The program should not stop until the user enters the /exit command.
5. If you encounter an empty line, do not output anything.