[master] 5/7 Error!
Description

Now, you need to consider the reaction of the calculator when users enter expressions in the wrong format. The program should only accept numbers, a plus + sign , a minus - sign , and two commands: /exit and /help. It cannot accept all other characters, and it is necessary to warn the user about this.

Objectives

1. The program should print Invalid expression in cases when the given expression has an invalid format. If a user enters an invalid command, the program must print Unknown command. All messages must be printed without quotes. The program must never throw an exception.
2. To handle incorrect input, you should remember that the user input that starts with / is a command; in other situations, it is an expression.
3. Do not forget to write methods to decompose your program.
4. Like before, the /help command should print information about your program. When the command /exit is entered, the program must print Bye! , and then stop.