[master] 6/7 Variables
Description:

In the beginning, we mentioned that the calculator would be able to store the results of previous calculations. Do you have any idea how to do that? Of course! This can be achieved by introducing variables. Storing results in variables and then operating on them at any time is a very convenient function.

Objectives:

Go by the following rules for variables:
1. We suppose that the name of a variable (identifier) can contain only Latin letters.
2. A variable can have a name consisting of more than one letter.
3. The case is also important; your program should be able to take both lowercase and uppercase variables. for example, z is not the same as Z.
4. The value can be an integer number or a value of another variable.
5. It should be possible to set a new value to an existing variable.
6. To print the value of a variable, you should just type its name.